refactor(logging): move runtime setup into logger module
